Bugfixing .NET Program.cs
When the parameters are parsed, it used " " to split the parameters. This currently causes a file named as such: "C:\Users\Linguz\My Files\etc.txt" to be split incorrectly. 
Currently, it would put "path" into paramName (as expected), but it would put "C:\Users\Linguz\My" into paramValue, which is an incomplete path. By using the Split overload, we can cap the number of splits into two: path/url (paramName) and then everything else.